backend/pyproject.toml
[tool.poetry]
name = "E-commerce backend"
version = "0.1.0"
description = ""
authors = ["Ilya Seleznev <seleznev.ilya.a@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.103.0"}
sqlalchemy = "^2.0.20"
asyncpg = "^0.28.0"
alembic = "^1.11.3"
gunicorn = "^21.2.0"
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
greenlet = "^2.0.2"
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
fastapi-cache2 = "^0.2.1"
redis = "^5.0.1"
celery = "^5.3.6"
flower = "^2.0.1"
sqladmin = "^0.16.0"
flake8 = "^7.0.0"
autoflake = "^2.2.1"
isort = "^5.13.2"
pyright = "^1.1.348"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
(src/tests|src/migrations)
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.autoflake]
check_diff = true
imports = ["sqlalchemy", "src", "fastapi", "sqladmin"]

[tool.pyright]
include = ["src"]
